requirements.txt
asgiref==3.8.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gitdb==4.0.10
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
platformdirs==4.2.2
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.9.0
Pympler==1.0.1
python-dateutil==2.8.2
python-pptx==1.0.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
six==1.16.0
smmap==5.0.0
sqlparse==0.5.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
virtualenv==20.26.3
watchdog==3.0.0
XlsxWriter==3.2.0
zipp==3.16.2
